// src/app/shared/enums/custom-media-type.enum.ts
export enum CustomMediaTypeEnum {
  EVENT_TRANSLATION = 'application/vnd+nova.event_translation+json',
  CHOICE_TRANSLATION = 'application/vnd+nova.choice_translation+json',
  DETAILED_EVENT = 'application/vnd+nova.detailed_event+json',
  EVENT_TITLE = 'application/vnd+nova.event_title+json',
  RESOURCE_WITH_TRANSLATIONS = 'application/vnd+nova.resource_translations+json',
  RESOURCE_NAME = 'application/vnd+nova.resource_name+json',
  LANGUAGE_WITH_AVAILABLE_ACTIONS = 'application/vnd+nova.language_with_available_actions+json'
}